Implemented a feature for duplication a cabin

+++ src/services/apiCabins.js
@@ -49,6 +49,8 @@ export async function createEditCabin(newCabin, id) {
         throw new Error("Cabin could not be created")
     }
 
+    if (hasImagePath) return data;
+
     const { error: storageError } = await supabase.storage.from('cabin-images').upload(imageName, newCabin.image)
     if (storageError) {
         await supabase.from('cabins').delete().eq('id', data.id);
